Refactoring
- Refactored folder structure
- Added pages directory
