+++ src/helper.ts
@@ -1,3 +1,4 @@
+import { TransactionCategories } from 'components/AddTransactionModal/TransactionCategoryInput/interface';
 import {
   DEBIT_TYPE,
 } from 'Constants';
@@ -20,4 +21,12 @@ export const isDebitTypeTransaction = (transaction: Transaction) => {
   // to handle the transactions where type debit or credit is not stored
   // adding undefined match also
   return transaction.type === DEBIT_TYPE || transaction.type === undefined;
+}
+
+export const checkTransactionCategoriesChanged = (data: TransactionCategories, storeTransactionCategories: TransactionCategories) => {
+  const { credit, debit } = storeTransactionCategories;
+  const { credit: dbCredit, debit: dbDebit } = data;
+  if (credit.length !== dbCredit.length) return true;
+  if (debit.length !== dbDebit.length) return true;
+  return false;
 }
