+++ src/pages/TransactionCategoriesPage/DisplayCategories/interface.ts
