+++ src/pages/TransactionCategoriesPage/DisplayCategories/styles.module.scss
