+++ src/pages/TransactionCategoriesPage/styles.module.scss
